+++ NEWS.md
@@ -1,3 +1,11 @@
+# normalize 0.1.2
+
+* Object attributes are preserved.
+
+* Added fallback `normalize.default()` method.
+
+* Small bug fixes.
+
 # normalize 0.1.1
 
 * Small bug fixes.